// src/api/swapi/types.ts
export interface IAllCharactersResponse {
  count: number;
  next: string | null;
  previous: string | null;
  results: IRawCharacter[];
}

export interface IAllCharactersData {
  count: number;
  next: string | null;
  previous: string | null;
  results: ICharacter[];
}

export interface IRawCharacter {
  birth_year: string;
  created: string;
  edited: string;
  eye_color: string;
  films: string[];
  gender: string;
  hair_color: string;
  height: string;
  homeworld: string;
  mass: string;
  name: string;
  skin_color: string;
  species: string[];
  starships: string[];
  url: string;
  vehicles: string[];
}

export interface ICharacter extends IRawCharacter {
  _id: string;
  _imageSrc: string;
}